+++ CourseSchedule/src/main/java/at/htl/courseschedule/controller/ParticipantRepository.java
@@ -12,6 +12,10 @@ public class ParticipantRepository implements Persistent<Participant> {
 
     @Override
     public void save(Participant participant) {
+        if (participant == null) {
+            return;
+        }
+
         if (participant.getId() == null) {
             insert(participant);
         }
@@ -22,6 +26,10 @@ public class ParticipantRepository implements Persistent<Participant> {
 
     @Override
     public void update(Participant participant) {
+        if (participant == null) {
+            return;
+        }
+
         try (Connection connection = dataSource.getConnection()) {
             String sql = "UPDATE CS_PARTICIPANT SET P_FIRST_NAME=?, P_LAST_NAME=?, P_YEAR_OF_BIRTH=?, P_PHONE_NR=?, P_EMAIL=? WHERE P_ID=?";
 
@@ -43,6 +51,10 @@ public class ParticipantRepository implements Persistent<Participant> {
 
     @Override
     public void insert(Participant participant) {
+        if (participant == null) {
+            return;
+        }
+
         try (Connection connection = dataSource.getConnection()) {
             String sql = "INSERT INTO CS_PARTICIPANT (P_FIRST_NAME, P_LAST_NAME, P_YEAR_OF_BIRTH, P_PHONE_NR, P_EMAIL) VALUES (?,?,?,?,?)";
 
@@ -71,6 +83,10 @@ public class ParticipantRepository implements Persistent<Participant> {
 
     @Override
     public void delete(Participant participant) {
+        if (participant == null) {
+            return;
+        }
+
         try (Connection connection = dataSource.getConnection()) {
             String sql = "DELETE FROM CS_PARTICIPANT WHERE P_ID=?";
 
@@ -116,7 +132,8 @@ public class ParticipantRepository implements Persistent<Participant> {
         Participant participant = null;
 
         try (Connection connection = dataSource.getConnection()) {
-            String sql = "SELECT * FROM CS_PARTICIPANT WHERE P_ID=?";
+            String sql = "SELECT P_ID, P_FIRST_NAME, P_LAST_NAME, P_YEAR_OF_BIRTH, P_PHONE_NR, P_EMAIL " +
+                    "FROM CS_PARTICIPANT WHERE P_ID=?";
 
             PreparedStatement statement = connection.prepareStatement(sql);
             statement.setLong(1, id);
